+++ HtmlRun.Common.Runtime/Native/INativeJSInstruction.cs
@@ -1,8 +1,6 @@
-using HtmlRun.Runtime.Interfaces;
-
 namespace HtmlRun.Runtime.Native;
 
-public interface INativeJSInstruction
+public interface INativeJSInstruction : INativeJSBaseInstruction
 {
   Delegate ToJSAction();
 }