+++ HtmlRun.Common/Enums/AppType.cs
@@ -1,0 +1,9 @@
+namespace HtmlRun.Common.Models;
+
+public enum AppType
+{
+  Unknown = 0,
+  Library = 1,
+  Console = 2,
+  WebServer = 3,
+}