+++ HtmlRun.Runtime/Compiler/HtmlRuntimeCompiler.cs
@@ -1,5 +1,7 @@
 using HtmlRun.Common.Models;
 using HtmlRun.Runtime.Compiler.BranchedInstructions;
+using HtmlRun.Runtime.Constants;
+using HtmlRun.Runtime.Factories;
 
 namespace HtmlRun.Runtime;
 
@@ -14,8 +16,57 @@ static class HtmlRuntimeCompiler
   {
     Parallel.ForEach(app.InstructionGroups, group =>
     {
+      AddGroupStartLabels(group);
+
+      EnsureReturns(group);
+
       CompileBranches(group);
+
+      AddGroupEndLabels(group);
     });
+
+    //  Prepare main group for the merge
+    var mainGroup = app.InstructionGroups.First(m => m.IsMain);
+    mainGroup.Instructions.Add(CallModelFactory.Goto("application-end"));
+
+    //  Merge and clear groups
+    foreach (var groupToMerge in app.LabeledGroups)
+    {
+      mainGroup.Instructions.AddRange(groupToMerge.Instructions);
+
+      groupToMerge.Instructions.Clear();
+    }
+
+    //  No los remuevo porque se usan para poder encontrar por label
+    // app.InstructionGroups.RemoveAll(m => !m.IsMain);
+
+    //  Add application end
+    mainGroup.Instructions.Add(CallModelFactory.Label("application-end"));
+  }
+
+  private static void AddGroupStartLabels(InstructionsGroup group)
+  {
+    string key = $"group-{group.Label}";
+    group.Instructions.Insert(0, CallModelFactory.Label(key));
+  }
+
+  private static void AddGroupEndLabels(InstructionsGroup group)
+  {
+    string key = $"end-group-{group.Label}";
+    group.Instructions.Add(CallModelFactory.Label(key));
+  }
+
+  private static void EnsureReturns(InstructionsGroup group)
+  {
+    if (!group.IsMain)
+    {
+      bool isReturn = group.Instructions.Last().FunctionName == BasicInstructionsSet.Return;
+
+      if (!isReturn)
+      {
+        group.Instructions.Add(CallModelFactory.Return());
+      }
+    }
   }
 
   private static void CompileBranches(InstructionsGroup group)