+++ HtmlRun.Runtime/Interfaces/IJumpReturn.cs
@@ -1,0 +1,5 @@
+namespace HtmlRun.Interfaces;
+
+public interface IJumpReturn
+{
+}