+++ HtmlRun.Runtime/Native/NativeJsNetDefinition.cs
@@ -1,0 +1,11 @@
+namespace HtmlRun.Runtime.Native;
+
+public class NativeJSNetDefinition : INativeJSDefinition
+{
+  public Delegate Delegate { get; set; }
+
+  public NativeJSNetDefinition(Delegate @delegate)
+  {
+    Delegate = @delegate;
+  }
+}