+++ HtmlRun.Runtime/RuntimeContext/InstructionPointer.cs
@@ -1,4 +1,5 @@
 using HtmlRun.Common.Models;
+using HtmlRun.Interfaces;
 using HtmlRun.Runtime.Interfaces;
 using HtmlRun.Runtime.RuntimeContext;
 
@@ -7,6 +8,8 @@ namespace HtmlRun.Runtime;
 class InstructionPointer
 {
   public int Position { get; private set; } = 0;
+
+  public Stack<IJumpWithMemory> CallStack { get; private set; } = new Stack<IJumpWithMemory>();
 
   public void MoveToNextPosition()
   {
@@ -19,6 +22,12 @@ class InstructionPointer
     {
       this.ApplyJump(jump, instructions);
     }
+    else if (cursorModification is IJumpReturn)
+    {
+      var call = this.CallStack.Pop() ?? throw new InvalidOperationException("Call stack is empty.");
+      
+      this.ApplyJump(new JumpToLine(call.CallPosition + 1), instructions);
+    }
     else
     {
       throw new InvalidCastException();
@@ -27,6 +36,12 @@ class InstructionPointer
 
   private void ApplyJump(JumpToLine jump, List<CallModel> instructions)
   {
+    if (jump is IJumpWithMemory jumpWithMemory)
+    {
+      jumpWithMemory.CallPosition = this.Position;
+      this.CallStack.Push(jumpWithMemory);
+    }
+
     if (jump.JumpType == JumpToLine.JumpTypeEnum.LineNumber)
     {
       this.Position = int.Parse(jump.Line!) - 1;