+++ HtmlRun.Tests/Runtime/HtmlRuntimeExamplesTests.cs
@@ -2,27 +2,24 @@ using HtmlRun.Common.Models;
 using HtmlRun.Interpreter.Interpreters;
 using HtmlRun.Runtime;
 using HtmlRun.Tests.Stubs;
-using HtmlRun.Tests.Stubs.Instructions;
 
-public class HtmlRuntimeExamplesTests: IDisposable
+public class HtmlRuntimeExamplesTests
 {
   private HtmlRuntime runtime;
+
   private SpiderInterpreter spider;
+
+  private List<string> logs;
 
   public HtmlRuntimeExamplesTests()
   {
-    LogCmd.Logs.Clear();
+    this.logs = new List<string>();
 
     this.runtime = new HtmlRuntime();
     this.runtime.RegisterBasicProviders();
-    this.runtime.RegisterProvider(new InstructionsProvider());
+    this.runtime.RegisterProvider(new InstructionsProvider(this.logs));
 
     this.spider = new SpiderInterpreter();
-  }
-
-  public void Dispose()
-  {
-    LogCmd.Logs.Clear();
   }
 
   [Fact]
@@ -32,8 +29,8 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Single(LogCmd.Logs);
-    Assert.Equal("Hello world", LogCmd.Logs[0]);
+    Assert.Single(this.logs);
+    Assert.Equal("Hello world", this.logs[0]);
   }
 
   [Fact]
@@ -41,10 +38,13 @@ public class HtmlRuntimeExamplesTests: IDisposable
   {
     var app = await this.ReadApp("02-simple_operation");
 
+    Assert.Equal("1.0.1", app.Version);
+    Assert.Equal(AppType.Console, app.Type);
+
     this.runtime.Run(app, null);
 
-    Assert.Single(LogCmd.Logs);
-    Assert.Equal("The result of 1 + 2 is 3", LogCmd.Logs[0]);
+    Assert.Single(this.logs);
+    Assert.Equal("The result of 1 + 2 is 3", this.logs[0]);
   }
 
   [Fact]
@@ -54,8 +54,8 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Single(LogCmd.Logs);
-    Assert.Equal("The result of Fibonacci(13) is 233", LogCmd.Logs[0]);
+    Assert.Single(this.logs);
+    Assert.Equal("The result of Fibonacci(13) is 233", this.logs[0]);
   }
 
   [Fact]
@@ -65,9 +65,9 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(2, LogCmd.Logs.Count);
-    Assert.Equal("3 > 2... Yep", LogCmd.Logs[0]);
-    Assert.Equal("1 > 2... Nope", LogCmd.Logs[1]);
+    Assert.Equal(2, this.logs.Count);
+    Assert.Equal("3 > 2... Yep", this.logs[0]);
+    Assert.Equal("1 > 2... Nope", this.logs[1]);
   }
 
   [Fact]
@@ -77,12 +77,12 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(5, LogCmd.Logs.Count);
-    Assert.Equal("Language: HyperText Machine Language", LogCmd.Logs[0]);
-    Assert.Equal("You have a new message: Hi there!", LogCmd.Logs[1]);
-    Assert.Equal("Reply: Hello", LogCmd.Logs[2]);
-    Assert.Contains("Current time: ", LogCmd.Logs[3]);
-    Assert.Contains("Next second value: ", LogCmd.Logs[4]);
+    Assert.Equal(5, this.logs.Count);
+    Assert.Equal("Language: HyperText Machine Language", this.logs[0]);
+    Assert.Equal("You have a new message: Hi there!", this.logs[1]);
+    Assert.Equal("Reply: Hello", this.logs[2]);
+    Assert.Contains("Current time: ", this.logs[3]);
+    Assert.Contains("Next second value: ", this.logs[4]);
   }
 
   [Fact]
@@ -92,11 +92,11 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    // Assert.Single(LogCmd.Logs);
-    Assert.Equal(3, LogCmd.Logs.Count);
-    Assert.Equal("Value: 0", LogCmd.Logs[0]);
-    Assert.Equal("Value (+1): 1", LogCmd.Logs[1]);
-    Assert.Equal("Value (-1): 0", LogCmd.Logs[2]);
+    // Assert.Single(this.logs);
+    Assert.Equal(3, this.logs.Count);
+    Assert.Equal("Value: 0", this.logs[0]);
+    Assert.Equal("Value (+1): 1", this.logs[1]);
+    Assert.Equal("Value (-1): 0", this.logs[2]);
   }
 
   [Fact]
@@ -106,13 +106,13 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(6, LogCmd.Logs.Count);
-    Assert.Equal("Current iteration: 0", LogCmd.Logs[0]);
-    Assert.Equal("Current iteration: 1", LogCmd.Logs[1]);
-    Assert.Equal("Current iteration: 2", LogCmd.Logs[2]);
-    Assert.Equal("Current iteration: 3", LogCmd.Logs[3]);
-    Assert.Equal("Current iteration: 4", LogCmd.Logs[4]);
-    Assert.Equal("Exit", LogCmd.Logs[5]);
+    Assert.Equal(6, this.logs.Count);
+    Assert.Equal("Current iteration: 0", this.logs[0]);
+    Assert.Equal("Current iteration: 1", this.logs[1]);
+    Assert.Equal("Current iteration: 2", this.logs[2]);
+    Assert.Equal("Current iteration: 3", this.logs[3]);
+    Assert.Equal("Current iteration: 4", this.logs[4]);
+    Assert.Equal("Exit", this.logs[5]);
   }
 
   [Fact]
@@ -122,8 +122,8 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Single(LogCmd.Logs);
-    Assert.Equal("Sleep was called without the namespace (Threading.Sleep).", LogCmd.Logs[0]);
+    Assert.Single(this.logs);
+    Assert.Equal("Sleep was called without the namespace (Threading.Sleep).", this.logs[0]);
   }
 
   [Fact]
@@ -133,8 +133,8 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Single(LogCmd.Logs);
-    Assert.Equal("3 > 2... Yep", LogCmd.Logs[0]);
+    Assert.Single(this.logs);
+    Assert.Equal("3 > 2... Yep", this.logs[0]);
   }
 
   [Fact]
@@ -144,10 +144,10 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(3, LogCmd.Logs.Count);
-    Assert.Equal("Arf arf!", LogCmd.Logs[0]);
-    Assert.Equal("Squeek!", LogCmd.Logs[1]);
-    Assert.Equal("Moo!", LogCmd.Logs[2]);
+    Assert.Equal(3, this.logs.Count);
+    Assert.Equal("Arf arf!", this.logs[0]);
+    Assert.Equal("Squeek!", this.logs[1]);
+    Assert.Equal("Moo!", this.logs[2]);
   }
 
   [Fact]
@@ -157,12 +157,12 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(5, LogCmd.Logs.Count);
-    Assert.Equal("Iteration: 0", LogCmd.Logs[0]);
-    Assert.Equal("Iteration: 1", LogCmd.Logs[1]);
-    Assert.Equal("Iteration: 2", LogCmd.Logs[2]);
-    Assert.Equal("Iteration: 3", LogCmd.Logs[3]);
-    Assert.Equal("Iteration: 4", LogCmd.Logs[4]);
+    Assert.Equal(5, this.logs.Count);
+    Assert.Equal("Iteration: 0", this.logs[0]);
+    Assert.Equal("Iteration: 1", this.logs[1]);
+    Assert.Equal("Iteration: 2", this.logs[2]);
+    Assert.Equal("Iteration: 3", this.logs[3]);
+    Assert.Equal("Iteration: 4", this.logs[4]);
   }
 
   [Fact]
@@ -172,13 +172,13 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(6, LogCmd.Logs.Count);
-    Assert.Equal("Do-While is working fine when condition is false", LogCmd.Logs[0]);
-    Assert.Equal("Iteration: 0", LogCmd.Logs[1]);
-    Assert.Equal("Iteration: 1", LogCmd.Logs[2]);
-    Assert.Equal("Iteration: 2", LogCmd.Logs[3]);
-    Assert.Equal("Iteration: 3", LogCmd.Logs[4]);
-    Assert.Equal("Iteration: 4", LogCmd.Logs[5]);
+    Assert.Equal("Do-While is working fine when condition is false", this.logs[0]);
+    Assert.Equal("Iteration: 0", this.logs[1]);
+    Assert.Equal("Iteration: 1", this.logs[2]);
+    Assert.Equal("Iteration: 2", this.logs[3]);
+    Assert.Equal("Iteration: 3", this.logs[4]);
+    Assert.Equal("Iteration: 4", this.logs[5]);
+    Assert.Equal(6, this.logs.Count);
   }
 
   [Fact]
@@ -188,12 +188,12 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(5, LogCmd.Logs.Count);
-    Assert.Equal("Iteration: 0", LogCmd.Logs[0]);
-    Assert.Equal("Iteration: 1", LogCmd.Logs[1]);
-    Assert.Equal("Iteration: 2", LogCmd.Logs[2]);
-    Assert.Equal("Iteration: 3", LogCmd.Logs[3]);
-    Assert.Equal("Iteration: 4", LogCmd.Logs[4]);
+    Assert.Equal(5, this.logs.Count);
+    Assert.Equal("Iteration: 0", this.logs[0]);
+    Assert.Equal("Iteration: 1", this.logs[1]);
+    Assert.Equal("Iteration: 2", this.logs[2]);
+    Assert.Equal("Iteration: 3", this.logs[3]);
+    Assert.Equal("Iteration: 4", this.logs[4]);
   }
 
   [Fact]
@@ -203,13 +203,53 @@ public class HtmlRuntimeExamplesTests: IDisposable
 
     this.runtime.Run(app, null);
 
-    Assert.Equal(16*16, LogCmd.Logs.Count);
+    Assert.Equal(16 * 16, this.logs.Count);
+  }
+
+  //  Partial comment. Will be implemented in another PR.
+  /*[Fact]
+  public async void HtmlRuntime_Example_18()
+  {
+    var app = await this.ReadApp("18-library_import");
+
+    this.runtime.Run(app, null);
+
+    Assert.Single(this.logs);
+    Assert.Equal("H", this.logs[0]);
+  }*/
+
+  [Fact]
+  public async void HtmlRuntime_Example_19()
+  {
+    var app = await this.ReadApp("19-call_stack");
+
+    this.runtime.Run(app, null);
+
+    Assert.Equal(4, this.logs.Count);
+    Assert.Equal("Begin", this.logs[0]);
+    Assert.Equal("Inside the function", this.logs[1]);
+    Assert.Equal("Closing app...", this.logs[2]);
+    Assert.Equal("Exit", this.logs[3]);
+  }
+
+
+  [Fact]
+  public async void HtmlRuntime_Example_20()
+  {
+    var app = await this.ReadApp("20-instruction_groups");
+
+    this.runtime.Run(app, null);
+
+    Assert.Equal(3, this.logs.Count);
+    Assert.Equal("Program started", this.logs[0]);
+    Assert.Equal("Info", this.logs[1]);
+    Assert.Equal("Program finished", this.logs[2]);
   }
 
   private async Task<AppModel> ReadApp(string exampleName)
   {
     string file = Environment.CurrentDirectory.Remove(Environment.CurrentDirectory.IndexOf("HtmlRun.Tests"));
-    var app = await this.spider.ParseString(File.ReadAllText($"{file}/Examples/{exampleName}.html"));
+    var app = await this.spider.ParseString($"{exampleName}.html", m => File.ReadAllText($"{file}/Examples/{m}"));
     Assert.NotNull(app);
     return app!;
   }